requirements.txt
Pillow
python-dotenv
argparse